Move InsufficientProbesException in ubic.gemma.core.analysis.preprocess.filter
